requirements.txt
accelerate==1.8.1
fastapi==0.116.0
pillow==11.3.0
python-multipart==0.0.20
torch==2.7.1
torchvision==0.22.1
transformers==4.53.1
uvicorn==0.35.0